+++ src/main/java/lotto/user/validation/ValidationPurchasePrice.java
@@ -7,9 +7,9 @@ import java.util.regex.Pattern;
 public enum ValidationPurchasePrice {
     CHECK_PRICE_INTEGER(price -> Pattern.matches("^[0-9]*$", price),
             "[ERROR] 구입금액은 자연수여야 합니다.", true),
-    CHECK_PRICE_NUMBER(money -> Integer.parseInt(money) % 1000 == 0,
+    CHECK_PRICE_NUMBER(money -> Long.parseLong(money) % 1000 == 0,
             "[ERROR] 구입 금액은 1000원 단위여야 합니다.", true),
-    CHECK_PRICE_POSITIVE(money -> Integer.parseInt(money) >= 0,
+    CHECK_PRICE_POSITIVE(money -> Long.parseLong(money) >= 0,
             "[ERROR] 금액은 양수여야 합니다.", true),
     NORMAL((numbers) -> (true), "", false);
